[frontend] Fix tensor list alias annotation
The current logic checks alias info before `[]` and after. If no alias info exists after `[]`, we overwrite the alias info before. This logic failed on argument like `Tensor(a!)[]`, dropping the alias info before `[]` on the floor. This PR adds a new alias info if it's missing after `[]`. This way we can keep the alias info before `[]`.

ghstack-source-id: 36478a993fb148c68811b723a5e8ca3ce12cf7ef
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83921